MLN-7315 | feat: add checks to integration test
